Building, Robot, Script and World created
The Robot is a car which obviously contains 4 wheels.

The Building is a house with 3 rooms. I wasn't able to add the doors.
When I tried to add any feature like a wall or a window, the
program closed automatically.

The Script is a c++ file that displays a welcome message in the terminal
when we open the world.

The world contains the Building, the Robot and the Script.